+++ R/vroom.R
@@ -29,7 +29,7 @@ NULL
 #'   either a character vector of types, `TRUE` or `FALSE`. See
 #'   [vroom_altrep()] for for full details.
 #' @param altrep_opts \Sexpr[results=rd, stage=render]{lifecycle::badge("deprecated")}
-#' @param show_col_specs Control showing the column specifications. If `TRUE`
+#' @param show_col_types Control showing the column specifications. If `TRUE`
 #'   column specifications are always show, if `FALSE` they are never shown. If
 #'   `NULL` (the default) they are shown only if an explicit specification is not
 #'   given to `col_types`.
@@ -110,7 +110,7 @@ vroom <- function(
   altrep_opts = deprecated(),
   num_threads = vroom_threads(),
   progress = vroom_progress(),
-  show_col_specs = NULL,
+  show_col_types = NULL,
   .name_repair = "unique"
   ) {
 
@@ -143,7 +143,7 @@ vroom <- function(
 
   col_select <- vroom_enquo(rlang::enquo(col_select))
 
-  has_spec <- !is.null(col_types)
+  has_col_types <- !is.null(col_types)
 
   col_types <- as.col_spec(col_types)
 
@@ -166,18 +166,18 @@ vroom <- function(
 
   out <- vroom_select(out, col_select, id)
 
-  if (should_show_col_spec(has_spec, show_col_specs)) {
-    show_col_specs(out, locale)
+  if (should_show_col_types(has_col_types, show_col_types)) {
+    show_col_types(out, locale)
   }
 
   out
 }
 
-should_show_col_spec <- function(has_spec, show_col_specs) {
-  if (is.null(show_col_specs)) {
-    return(isTRUE(!has_spec))
+should_show_col_spec <- function(has_col_types, show_col_types) {
+  if (is.null(show_col_types)) {
+    return(isTRUE(!has_col_types))
   }
-  isTRUE(show_col_specs)
+  isTRUE(show_col_types)
 }
 
 make_names <- function(x, len) {